Config download end-to-end flow wiring
