+++ crates/bin/c8y_configuration_poc/src/c8y_http_proxy/actor.rs
@@ -100,6 +100,7 @@ impl C8YHttpProxyActor {
                 C8YRestRequest::C8yUpdateSoftwareListResponse(_) => {}
                 C8YRestRequest::UploadLogBinary(_) => {}
                 C8YRestRequest::UploadConfigFile(_) => {}
+                C8YRestRequest::DownloadFile(_) => {}
             }
         }
         Ok(())
